Update AZ400_M03_Configuring_Agent_Pools_and_Understanding_Pipeline_Styles.md
Demonstrate proper and necessary code indentation in yaml files by having the user replace the section starting with `pool:`. It isn't necessary for the instructions but does present the opportunity for learning and help prevent an unexpected error. I discovered it created an issue when the insertion tool that added the lines for me did not maintain the leading spaces, therefore the build failed. I also reformatted the text a bit because it appeared to be broken.